Removed VideoInfo (inlined to Video)
VideoInfo is not needed anymore after switching to OptimizedVideo
